Move Participant class into its own file
