+++ app/logic/Calendar/JSON/JSONEvent.ts
@@ -1,4 +1,5 @@
-import { Event, Participant } from "../Event";
+import { Event } from "../Event";
+import { Participant } from "../Participant";
 import { RecurrenceRule } from "../RecurrenceRule";
 import { appGlobal } from "../../app";
 import { sanitize } from "../../../../lib/util/sanitizeDatatypes";
